Renamed module name from upgrade to hotfix
