+++ x/hotfix/expected/keeper.go
@@ -3,7 +3,6 @@ package expected
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
-	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
 )
 
 type AccountKeeper interface {
@@ -13,8 +12,4 @@ type AccountKeeper interface {
 
 type BankKeeper interface {
 	GetAllBalances(ctx sdk.Context, address sdk.AccAddress) sdk.Coins
-}
-
-type UpgradeKeeper interface {
-	ScheduleUpgrade(ctx sdk.Context, plan upgradetypes.Plan) error
 }
