+++ x/hotfix/hotfix-1/keys.go
@@ -1,0 +1,6 @@
+package hotfix1
+
+const (
+	Name   = "hotfix-1"
+	Height = 940000
+)
