--- x/upgrade/upgrade-1/keys.go
@@ -1,6 +1,0 @@
-package upgrade1
-
-const (
-	Name   = "upgrade-1"
-	Height = 940000
-)
